+++ Cargo.toml
@@ -135,8 +135,6 @@ mimalloc         = { git = "https://github.com/utooland/mimalloc_rust.git" }
 parcel_selectors = { git = "https://github.com/utooland/lightningcss", branch = "support-css-module-global" }
 tokio            = { package = "tokio", git = "https://github.com/utooland/tokio.git" }
 tokio-util       = { package = "tokio-util", git = "https://github.com/utooland/tokio.git" }
-opfs-project     = { path = "../opfs-project" }
-tokio-fs-ext     = { path = "../tokio-fs-ext" }
 virtue           = { git = "https://github.com/bgw/virtue.git", branch = "bgw/fix-generic-default-parsing" }
 
 [profile.release]
