+++ packages/utoo-web/src/utoo/index.d.ts
@@ -194,18 +194,13 @@ export function workerCreated(worker_id: number): void;
 export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembly.Module;
 
 export interface InitOutput {
-    readonly ERR_ABORT: () => [number, number];
-    readonly ERR_INVALID_STATE: () => [number, number];
-    readonly ERR_KEY_ALREADY_EXISTS: () => [number, number];
-    readonly ERR_NOT_ALLOWED: () => [number, number];
-    readonly ERR_NOT_FOUND: () => [number, number];
-    readonly ERR_NO_MODIFICATION_ALLOWED: () => [number, number];
-    readonly ERR_QUOTA_EXCEEDED: () => [number, number];
-    readonly ERR_TYPE_MISMATCH: () => [number, number];
-    readonly getWasmMemory: () => any;
-    readonly getWasmModule: () => any;
     readonly initLogFilter: (a: number, b: number) => void;
     readonly init_pack: () => void;
+    readonly getWasmMemory: () => any;
+    readonly getWasmModule: () => any;
+    readonly __wbg_roottask_free: (a: number, b: number) => void;
+    readonly registerWorkerScheduler: (a: any, b: any) => void;
+    readonly workerCreated: (a: number) => void;
     readonly __wbg_project_free: (a: number, b: number) => void;
     readonly project_build: () => any;
     readonly project_cwd: () => [number, number];
@@ -219,6 +214,14 @@ export interface InitOutput {
     readonly project_sigMd5: (a: any) => any;
     readonly project_updateInfoSubscribe: (a: number, b: any) => void;
     readonly project_writeAllToDisk: (a: any) => void;
+    readonly ERR_ABORT: () => [number, number];
+    readonly ERR_INVALID_STATE: () => [number, number];
+    readonly ERR_KEY_ALREADY_EXISTS: () => [number, number];
+    readonly ERR_NOT_ALLOWED: () => [number, number];
+    readonly ERR_NOT_FOUND: () => [number, number];
+    readonly ERR_NO_MODIFICATION_ALLOWED: () => [number, number];
+    readonly ERR_QUOTA_EXCEEDED: () => [number, number];
+    readonly ERR_TYPE_MISMATCH: () => [number, number];
     readonly __wbg_direntry_free: (a: number, b: number) => void;
     readonly __wbg_fs_free: (a: number, b: number) => void;
     readonly __wbg_get_direntry_name: (a: number) => [number, number];
@@ -250,9 +253,6 @@ export interface InitOutput {
     readonly fs_write: (a: number, b: number, c: any) => any;
     readonly fs_writeString: (a: number, b: number, c: number, d: number) => any;
     readonly fs_writeSync: (a: number, b: number, c: any) => [number, number];
-    readonly __wbg_roottask_free: (a: number, b: number) => void;
-    readonly registerWorkerScheduler: (a: any, b: any) => void;
-    readonly workerCreated: (a: number) => void;
     readonly rust_mi_get_default_heap: () => number;
     readonly rust_mi_get_thread_id: () => number;
     readonly rust_mi_set_default_heap: (a: number) => void;
@@ -278,16 +278,16 @@ export interface InitOutput {
     readonly __wbg_createsyncaccesshandleoptions_free: (a: number, b: number) => void;
     readonly wasm_thread_entry_point: (a: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_core_e1885f5e8fe25124___ops__function__FnMut_____Output_______: (a: number, b: number) => void;
-    readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_core_e1885f5e8fe25124___ops__function__Fn__js_sys_d20acbc549183cf6___Array____Output_______: (a: number, b: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_core_e1885f5e8fe25124___ops__function__FnMut__wasm_bindgen_f95c71b28f29bc9d___JsValue____Output_______: (a: number, b: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_core_e1885f5e8fe25124___ops__function__FnMut__web_sys_27d1e2f4b4c95162___features__gen_MessageEvent__MessageEvent____Output_______: (a: number, b: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_core_e1885f5e8fe25124___ops__function__FnMut_____Output________1_: (a: number, b: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_for__a__core_e1885f5e8fe25124___ops__function__FnMut____a_web_sys_27d1e2f4b4c95162___features__gen_MessageEvent__MessageEvent____Output_______: (a: number, b: number) => void;
+    readonly wasm_bindgen_f95c71b28f29bc9d___closure__destroy___dyn_core_e1885f5e8fe25124___ops__function__Fn__js_sys_d20acbc549183cf6___Array____Output_______: (a: number, b: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke___js_sys_d20acbc549183cf6___Function__js_sys_d20acbc549183cf6___Function_____: (a: number, b: number, c: any, d: any) => void;
-    readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke___js_sys_d20acbc549183cf6___Array_____: (a: number, b: number, c: any) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke___wasm_bindgen_f95c71b28f29bc9d___JsValue_____: (a: number, b: number, c: any) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke___web_sys_27d1e2f4b4c95162___features__gen_MessageEvent__MessageEvent_____: (a: number, b: number, c: any) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures________invoke___web_sys_27d1e2f4b4c95162___features__gen_MessageEvent__MessageEvent_____: (a: number, b: number, c: any) => void;
+    readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke___js_sys_d20acbc549183cf6___Array_____: (a: number, b: number, c: any) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke______: (a: number, b: number) => void;
     readonly wasm_bindgen_f95c71b28f29bc9d___convert__closures_____invoke_______1_: (a: number, b: number) => void;
     readonly memory: WebAssembly.Memory;
